Fixed subroutine for searching of execution command
Without this fix colorquestasim do not work in Linux.

Main target of colorquestasim is to run the QuestaSim 'vlog' command.
On Linux, <PathInEnv>/vlog is a link to '../vco'. But 'vlog' can not be
run via this link, to run 'vlog' and other command need execute they
links ('vlog', 'vopt', 'vsim'). But old version launch 'vco'.